refactor: Car 인스턴스 변수 name을 Name 클래스로 포장
