+++ src/main/java/racingcar/model/Car.java
@@ -2,9 +2,8 @@ package racingcar.model;
 
 public class Car {
     private static final int MOVE_POINT_NUMBER = 4;
-    private static final int NAME_LEN = 5;
 
-    private final String name;
+    private final Name name;
     private int distance;
 
     public Car(final String name) {
@@ -12,15 +11,8 @@ public class Car {
     }
 
     public Car(final String name, final int distance) {
-        validateName(name);
-        this.name = name;
+        this.name = new Name(name);
         this.distance = distance;
-    }
-
-    private void validateName(final String name) {
-        if (name.length() > NAME_LEN) {
-            throw new IllegalArgumentException("[Error] 자동차 이름은 5자 이하여야 합니다.\n");
-        }
     }
 
     public void move(final int number) {
@@ -38,7 +30,7 @@ public class Car {
     }
 
     public String getName() {
-        return this.name;
+        return this.name.getName();
     }
 
     public int getDistance() {
